+++ src/main/java/com/sparta/goodbite/domain/operatinghour/dto/CreateOperatingHourRequestDto.java
@@ -1,8 +1,8 @@
 package com.sparta.goodbite.domain.operatinghour.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.sparta.goodbite.domain.operatinghour.dto.validation.contraint.LocalTimeFormatConstraint;
-import com.sparta.goodbite.domain.operatinghour.dto.validation.contraint.OpenTimeBeforeCloseTimeConstraint;
+import com.sparta.goodbite.common.validation.constraint.LocalTimeFormatConstraint;
+import com.sparta.goodbite.common.validation.constraint.OpenTimeBeforeCloseTimeConstraint;
 import com.sparta.goodbite.domain.operatinghour.entity.OperatingHour;
 import com.sparta.goodbite.domain.operatinghour.enums.DayOfWeek;
 import com.sparta.goodbite.domain.restaurant.entity.Restaurant;
