+++ src/main/java/com/sparta/goodbite/domain/review/dto/UpdateReviewRequestDto.java
@@ -1,6 +1,6 @@
 package com.sparta.goodbite.domain.review.dto;
 
-import com.sparta.goodbite.domain.review.dto.validation.constraint.RatingConstraint;
+import com.sparta.goodbite.common.validation.constraint.RatingConstraint;
 import lombok.Getter;
 
 @Getter
